requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.2
django-cors-headers==3.0.2
django-filter==2.1.0
djangorestframework==3.9.4
idna==2.8
Markdown==3.1.1
mysqlclient==1.4.2.post1
Pillow==6.0.0
psycopg2==2.8.2
pycparser==2.19
PyJWT==1.7.1
python-decouple==3.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
